pyproject.toml
[project]
name = "ioavstats"
version = "25.01.15"
authors = [
    { name = "Team IO-Aero", email = "info@io-aero.com" },
]
maintainers = [
    { name = "Team IO-Aero", email = "info@io-aero.ch" },
]
description = "Aviation Accident Statistics"
readme = "README.md"
keywords = [
    "accidents",
    "application",
    "aviation",
    "ntsb",
    "statistics",
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Console",
    "Intended Audience :: Science/Research",
    "License :: Other/Proprietary License",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
    "Typing :: Typed",
]
requires-python = ">=3.10"
dependencies = [
    "altair<=5",
    "haversine>=2.9.0",
    "iocommon>=1.2.72",
    "lat_lon_parser>=1.3.1",
    "mlxtend>=0.23.3",
    "numpy>=1.26.4",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "plotly>=5.24.1",
    "protobuf>=3.20.3",
    "psycopg2-binary>=3.2.3",
    "pyeclat>=1.0.2",
    "pyodbc>=5.2.0",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.37",
    "streamlit>=1.41.1",
    "streamlit-pandas-profiling>=0.1.3", # not yet available for 'make docs'
    "tomli>=2.2.1",
    "virtualenv>=20.28.0",
    "xlrd>=2.0.1",
    "ydata-profiling>=4.12.1",
]

[project.urls]
"Bug Tracker" = "https://github.com/io-aero/io-avstats/issues"
Homepage = "https://github.com/io-aero/io-avstats"

[tool.bandit]
skips = [
    "B101",
	"B307",
	"B404",
	"B603",
	"B608",
]

[tool.docformatter]
black = true
blank = true
recursive = true
wrap-descriptions = 100
wrap-summaries = 100

[tool.pytest.ini_options]
console_output_style = "count"
markers = [
    "issue",
]
norecursedirs = "./tests/helpers"
pythonpath = "./ioavstats/"
testpaths = "./tests/"

[tool.ruff]
exclude = [
    ".git",
    ".history",
    ".idea",
    ".mypy_cache",
    ".pytest_cache",
    ".vscode",
    "build",
    "ioavstats.egg-info",
    "temp",
]
line-length = 100
indent-width = 4
target-version = "py312"
extend-ignore = [
    "E501",
]

[tool.ruff.format]
docstring-code-format = false
docstring-code-line-length = "dynamic"
indent-style = "space"
line-ending = "auto"
quote-style = "double"
skip-magic-trailing-comma = false

[tool.ruff.lint]
select = [
    "ALL",
]
ignore = [
    "C901",
    "D211",
    "D213",
    "D407",
    "PGH003",
    "PLR0911",
    "PLR0912",
    "PLR0913",
    "PLR0915",
    "PLR2004",
    "PLW0603",
    "S101",
    "S603",
    "S608",
]
fixable = [
    "ALL",
]
unfixable = []
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.lint.per-file-ignores]
"docs/source/conf.py" = ["D203"]
"ioavstats/db_ddl_base.py" = ["E501"]
"ioavstats/db_dml_corr.py" = ["E501"]
"ioavstats/db_dml_msaccess.py" = ["E501"]
"ioavstats/Menu.py" = ["E501"]
"ioavstats/pages/Association_Rule_Analysis.py" = ["E501"]
"ioavstats/pages/Aviation_Event_Analysis.py" = ["E501"]
"ioavstats/pages/Database_Profiling.py" = ["E501"]
"ioavstats/user_guide.py" = ["E501"]

[tool.vulture]
make_whitelist = true
min_confidence = 100
